containerized-app/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic
requests
